# dist/swagger/swagger.json
{
    "components": {
...
        "schemas": {
...
            "IApp": {
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "icon": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "tags": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array"
                    },
                    "categoryName": {
                        "type": "string"
                    },
                    "groupName": {
                        "type": "string"
                    },
                    "hasViewer": {
                        "type": "boolean"
                    },
                    "documentationLink": {
                        "type": "string"
                    },
                    "packageName": {
                        "type": "string"
                    },
                    "isExternalApp": {
                        "type": "boolean"
                    },
                    "link": {
                        "type": "string"
                    }
                },
                "required": [
                    "name",
                    "icon",
                    "description",
                    "tags",
                    "categoryName",
                    "groupName"
                ],
                "type": "object",
                "additionalProperties": {}
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "info": {
        "title": "BOS config admin API routes",
        "version": "1.0.0",
        "description": "Welcome to the reference documentation for Portofolio asset manager (PAM) REST API! </br></br>To learn about the common use cases and concept this REST APIs, check out our resource center  (<a>https://resourcecenter.fr.spinalcom.com/ressources-développeur</a>).</br></br>In addition to the API Reference, we also provide complete documentation for adminnistration API, that need to install and implement Spinalcore Digital Twin Operating System on their built environment.",
        "license": {
            "name": "ISC"
        },
        "contact": {
            "name": "Spinalcom",
            "email": "developers@spinalcom.com"
        }
    },
    "openapi": "3.0.0",
    "paths": {
...
        "/api/v1/pam/create_admin_app": {
            "post": {
                "operationId": "CreateAdminApp",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        {
                                            "$ref": "#/components/schemas/IApp"
                                        },
                                        {
                                            "properties": {
                                                "message": {
                                                    "type": "string"
                                                }
                                            },
                                            "required": [
                                                "message"
                                            ],
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Applications"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/IApp"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/pam/create_portofolio_app": {
            "post": {
                "operationId": "CreatePortofolioApp",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        {
                                            "$ref": "#/components/schemas/IApp"
                                        },
                                        {
                                            "properties": {
                                                "message": {
                                                    "type": "string"
                                                }
                                            },
                                            "required": [
                                                "message"
                                            ],
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Applications"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/IApp"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/pam/create_building_app": {
            "post": {
                "operationId": "CreateBuildingApp",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        {
                                            "$ref": "#/components/schemas/IApp"
                                        },
                                        {
                                            "properties": {
                                                "message": {
                                                    "type": "string"
                                                }
                                            },
                                            "required": [
                                                "message"
                                            ],
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Applications"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/IApp"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/pam/get_all_admin_apps": {
            "get": {
                "operationId": "GetAllAdminApps",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        {
                                            "items": {
                                                "$ref": "#/components/schemas/IApp"
                                            },
                                            "type": "array"
                                        },
                                        {
                                            "properties": {
                                                "message": {
                                                    "type": "string"
                                                }
                                            },
                                            "required": [
                                                "message"
                                            ],
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Applications"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": []
            }
        },
        "/api/v1/pam/get_all_portofolio_apps": {
            "get": {
                "operationId": "GetAllPortofolioApps",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        {
                                            "items": {
                                                "$ref": "#/components/schemas/IApp"
                                            },
                                            "type": "array"
                                        },
                                        {
                                            "properties": {
                                                "message": {
                                                    "type": "string"
                                                }
                                            },
                                            "required": [
                                                "message"
                                            ],
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Applications"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": []
            }
        },
...
    },
    "servers": [
        {
            "url": "/"
        }
    ]
}